% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_exploratory.R
\name{update_exploratory}
\alias{update_exploratory}
\title{Update the package 'exploratory'}
\usage{
update_exploratory(confirm = TRUE)
}
\arguments{
\item{confirm}{logical. If \code{confirm = TRUE}, the user will
need to confirm the update. If \code{confirm = FALSE}, the confirmation
step will be skipped. By default, \code{confirm = TRUE}.}
}
\value{
there will be no output from this function. Rather, executing
this function will update the current 'exploratory' package by installing
the most recent version of the package from GitHub.
}
\description{
Updates the current package 'exploratory' by installing the
most recent version of the package from GitHub
}
\examples{
\dontrun{
if (interactive()) {update_exploratory()}
}

}
